% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_latest_candidate_contributions.R
\name{download_latest_candidate_contributions}
\alias{download_latest_candidate_contributions}
\title{Download latest candidate contributions by candidate ID}
\usage{
download_latest_candidate_contributions(
  id,
  start_date = as.Date("2020-02-01"),
  boston_zips = known_boston_zipcodes
)
}
\arguments{
\item{id}{character string with candidate MA OCPF ID}

\item{start_date}{string in yyyy-mm-dd format or date object for contributions on or after a certain date, defaults to NULL (all candidate data downloads; this function filters it)}
}
\value{
data frame with contribution information
}
\description{
Download latest candidate contributions by candidate ID
}
